+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 CHANGELOG
 =========
 
-1.10.0
-------------------
+1.10.0 (2019-12-19)
+-------------------
 
 * IMPORTANT: Java 8 is now required. If you need Java 7 support, please
   continue using 1.9.0.